Refactor logger setup and main function extraction
Move logger and argument parsing functions to `config/utils.py` and refactor `if __name__ == '__main__'` section into a main function.

* Add `setup_logger` and `parse_arguments` functions to `config/utils.py`.
* Import `setup_logger` and `parse_arguments` from `config/utils.py` in `analysis/dataFrameAnalysis.py`.
* Extract logic in `if __name__ == '__main__'` section into a new `main` function in `analysis/dataFrameAnalysis.py`.
* Call the new `main` function in the `if __name__ == '__main__'` section in `analysis/dataFrameAnalysis.py`.
* Pass `args` as a parameter to `DataFrameAnalyzer` and `run_analysis` in `analysis/dataFrameAnalysis.py`.
* Import `setup_logger` from `config/utils.py` in `analysis/example_analysis.py`.
* Remove logger attribute from `UserFunctions` class and update it to use imported logger in `analysis/example_analysis.py`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ZGerber/taAnalysis?shareId=XXXX-XXXX-XXXX-XXXX).
